+++ common/Utils.lua
@@ -1,0 +1,49 @@
+local addonName, EXT = ...
+
+local L = EXT.localization
+
+local Utils = {}
+
+-----------------------
+--- Helper Funtions ---
+-----------------------
+
+---------------------
+--- Main Funtions ---
+---------------------
+
+function Utils:PrintDebug(msg)
+    if EXT.data.options["debug-mode"] then
+        local notfound = true
+
+        for i = 1, NUM_CHAT_WINDOWS do
+            local name, _, _, _, _, _, shown, locked, docked, uni = GetChatWindowInfo(i)
+
+            if name == "Debug" and docked ~= nil then
+                _G['ChatFrame' .. i]:AddMessage(ORANGE_FONT_COLOR:WrapTextInColorCode(addonName .. " (Debug): ")  .. msg)
+                notfound = false
+                break
+            end
+        end
+
+        if notfound then
+            DEFAULT_CHAT_FRAME:AddMessage(ORANGE_FONT_COLOR:WrapTextInColorCode(addonName .. " (Debug): ")  .. msg)
+        end
+	end
+end
+
+function Utils:PrintMessage(msg)
+    DEFAULT_CHAT_FRAME:AddMessage(NORMAL_FONT_COLOR:WrapTextInColorCode(addonName .. ": ") .. msg)
+end
+
+function Utils:InitializeDatabase()
+    -- Options
+    if (not Expositum_Options) then
+        Expositum_Options = {}
+    end
+
+    EXT.data = {}
+    EXT.data.options = Expositum_Options
+end
+
+EXT.utils = Utils