+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Expositum/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Expositum/releases)
 
+- Added: Options menu to customize the tooltip
+- Added: Language support
 - Minor code adjustments